backend/azure/pyproject.toml
[project]
name = "azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "azure-identity",
    "azure-mgmt-billing",
    "azure-mgmt-billingbenefits",
    "azure-mgmt-commerce",
    "azure-mgmt-consumption",
    "azure-mgmt-costmanagement",
    "azure-mgmt-reservations",
    "sqlalchemy",
    "aiomysql",
    "cryptography",
    "python-dotenv",
    "pydantic-settings",
    "six",
    "greenlet",
]
